requirements.txt
Django>=1.10,<2.0
django-crispy-forms==1.6.1
django-hitcount==1.3.0
wagtail>=1.12.1,<2.0
longclaw>=0.2.0
